# Conflicts:
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RASI/FileRW.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RASI/RasiExecutor.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RASI/RasiParser.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RASI/TeamRasiCommands.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RobotDrivers/IntakeController.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RobotDrivers/MecanumCont.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RobotDrivers/ServoPositions.json
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utilities/Geometry/Arc.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/MecanumTest.java

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utilities/Geometry/Point.java
@@ -31,7 +31,7 @@ public class Point{
     public String toString(){
         return "------------\nPoint\nX:"+this.xCord+"\nY:"+this.yCord+"\n------------\n";
     }
-    public boolean AreSame(Point compare, float tolerance){
+    public boolean AreSame(Point compare, double tolerance){
         if (Math.abs(this.xCord - compare.xCord) < tolerance && Math.abs(this.yCord - compare.yCord) < tolerance){
             return true;
         }
